Add basic data collection tied to two buttons
